+++ packages/api/src/feature/club/club.module.ts
@@ -4,8 +4,9 @@ import { DrizzleModule } from "src/drizzle/drizzle.module";
 
 import UserModule from "../user/user.module";
 
-import ClubDelegateController from "./controller/club-delegate.controller";
 import { ClubController } from "./controller/club.controller";
+import ClubDelegateController from "./delegate/delegate.controller";
+import ClubDelegateService from "./delegate/delegate.service";
 import { ClubDelegateDRepository } from "./repository/club.club-delegate-d.repository";
 import { ClubRoomTRepository } from "./repository/club.club-room-t.repository";
 import ClubStudentTRepository from "./repository/club.club-student-t.repository";
@@ -15,11 +16,11 @@ import { ClubGetStudentClubBrief } from "./repository/club.get-student-club-brie
 import { ClubPutStudentClubBrief } from "./repository/club.put-student-club-brief";
 import ClubRepository from "./repository/club.repository";
 import SemesterDRepository from "./repository/club.semester-d.repository";
-import ClubDelegateService from "./service/club-delegate.service";
 
 import ClubPublicService from "./service/club.public.service";
 import { ClubService } from "./service/club.service";
 
+// TODO: delegate 관련 모듈화하기
 @Module({
   imports: [DrizzleModule, UserModule],
   controllers: [ClubController, ClubDelegateController],
