+++ packages/api/src/feature/club/delegate/delegate.service.ts
@@ -5,11 +5,10 @@ import { ClubDelegateEnum } from "@sparcs-clubs/interface/common/enum/club.enum"
 import logger from "@sparcs-clubs/api/common/util/logger";
 import { getKSTDate } from "@sparcs-clubs/api/common/util/util";
 
+import ClubPublicService from "@sparcs-clubs/api/feature/club/service/club.public.service";
 import UserPublicService from "@sparcs-clubs/api/feature/user/service/user.public.service";
 
 import { ClubDelegateDRepository } from "../repository/club.club-delegate-d.repository";
-
-import ClubPublicService from "./club.public.service";
 
 import type {
   ApiClb006RequestParam,
@@ -84,39 +83,47 @@ export default class ClubDelegateService {
    * @param clubId 동아리 Id
    * @param clubDelegateEnumId 대표자 지위 Id
    *
-   * @description 동아리 대표자의 변경을 적용합니다.
+   * @description putStudentClubDelegate의 서비스 진입점입니다.
+   * 동아리 대표자의 변경을 적용합니다.
    */
   async putStudentClubDelegate(param: {
     studentId: number;
     targetStudentId: number;
     clubId: number;
     clubDelegateEnumId: number;
-  }) {
+  }): Promise<void> {
     const currentDelegates =
       await this.clubDelegateDRepository.findDelegateByClubId(param.clubId);
 
     // clubDelegateEnumId가 대표자일 경우, 신청자가 현재 동아리의 대표자가 맞는지 검사합니다.
     // clubDelegateEnumId가 대의원일 경우, 신청자가 현재 동아리의 대의원이 맞는지 검사합니다.
+    // 24.09.24 변경사항: 동아리 대표자만 사용 가능하도록 변경합니다.
     const studentStatus = currentDelegates.find(
       e => e.studentId === param.studentId,
     );
-    if (
-      studentStatus === undefined ||
-      (param.clubDelegateEnumId === ClubDelegateEnum.Representative &&
-        studentStatus.ClubDelegateEnumId !== ClubDelegateEnum.Representative)
-    )
+    // if (
+    //   studentStatus === undefined ||
+    //   (param.clubDelegateEnumId === ClubDelegateEnum.Representative &&
+    //     studentStatus.ClubDelegateEnumId !== ClubDelegateEnum.Representative)
+    // )
+    //   throw new HttpException(
+    //     "This api is allowed for delegates",
+    //     HttpStatus.FORBIDDEN,
+    //   );
+    if (studentStatus.ClubDelegateEnumId !== ClubDelegateEnum.Representative)
       throw new HttpException(
-        "This api is allowed for delegates",
+        "This api is allowed for the club representative",
         HttpStatus.FORBIDDEN,
       );
 
     // 신청자가 이미 다른 요청을 생성하지 않았는지 검사합니다.
+    // 대의원 변경의 경우 요청을 생성하지 않고 바로 변경되기에 항상 통과해야 합니다.
     const requests =
       await this.clubDelegateDRepository.findDelegateChangeRequestByPrevStudentId(
         { studentId: param.studentId },
       );
     if (requests.length > 1)
-      throw new HttpException("unreahable", HttpStatus.INTERNAL_SERVER_ERROR);
+      throw new HttpException("unreachable", HttpStatus.INTERNAL_SERVER_ERROR);
     if (requests.length === 1)
       throw new HttpException(
         "You already made a request",
@@ -149,18 +156,46 @@ export default class ClubDelegateService {
         HttpStatus.BAD_REQUEST,
       );
 
+    // 대표자 변경 신청의 경우 targetStudentId가 undefined인지 검사합니다.
     if (
-      !(await this.clubDelegateDRepository.insertClubDelegateChangeRequest({
-        clubId: param.clubId,
-        clubDelegateEnumId: param.clubDelegateEnumId,
-        studentId: param.studentId,
-        targetStudentId: param.targetStudentId,
-      }))
+      param.clubDelegateEnumId === ClubDelegateEnum.Representative &&
+      param.targetStudentId === undefined
     )
       throw new HttpException(
-        "Failed to insert request",
-        HttpStatus.INTERNAL_SERVER_ERROR,
+        "representative can;t be empty",
+        HttpStatus.BAD_REQUEST,
       );
+
+    // 대의원의 경우 즉시 변경하고, 대표자의 경우 변경 요청을 생성합니다.
+    switch (param.clubDelegateEnumId) {
+      case ClubDelegateEnum.Representative:
+        if (
+          !(await this.clubDelegateDRepository.insertClubDelegateChangeRequest({
+            clubId: param.clubId,
+            clubDelegateEnumId: param.clubDelegateEnumId,
+            studentId: param.studentId,
+            targetStudentId: param.targetStudentId,
+          }))
+        )
+          throw new HttpException(
+            "Failed to insert request",
+            HttpStatus.INTERNAL_SERVER_ERROR,
+          );
+        break;
+      default:
+        if (
+          !this.clubDelegateDRepository.updateDelegate({
+            clubId: param.clubId,
+            clubDelegateEnumId: param.clubDelegateEnumId,
+            studentId: param.studentId,
+          })
+        )
+          throw new HttpException(
+            "Failed to change delegate",
+            HttpStatus.INTERNAL_SERVER_ERROR,
+          );
+        break;
+    }
   }
 
   /**
